+++ src/regex.js
@@ -6,7 +6,9 @@ const regexMd = {
     "`": /`([^`]+)`/g,
     "#": /(#{1,6})\s?(.+)/g,
     "  ": /  \n/g,
-    "\n": /([^\n])(?!\n*<\/?h)\n([^\n])/g
+    "\n": /([^\n])(?!\n*<\/?h)\n([^\n])/g,
+    "p": /^(?!<h[1-6]>)([^\n]+)\n/gm,
+    "blockquote": /^>(.*(?:\n(?!>).*?)*)/gm
 }
 
 export default regexMd;
