+++ src/utils.js
@@ -1,0 +1,7 @@
+const pipe = functions => data => {
+  return functions.reduce(
+    (value, func) => func(value), data
+  );
+};
+
+export default pipe;
